+++ Clip Hold/App/Clip_HoldApp.swift
@@ -63,6 +63,7 @@ struct ClipHoldApp: App {
     @AppStorage("textOnlyQuickPaste") var textOnlyQuickPaste: Bool = false
     @AppStorage("showColorCodeIcon") var showColorCodeIcon: Bool = false
     @AppStorage("showCharacterCount") var showCharacterCount: Bool = false
+    @AppStorage("showInvisibleCharacters") var showInvisibleCharacters: Bool = false
     
     @StateObject var standardPhraseManager = StandardPhraseManager.shared
     @StateObject var clipboardManager = ClipboardManager.shared
@@ -195,7 +196,12 @@ struct ClipHoldApp: App {
                     let shortcutName = element.shortcutName
                     
                     let displayText: String = {
-                        let displayContent = phrase.title.replacingNewlinesWithSpaces()
+                        let displayContent: String
+                        if showInvisibleCharacters {
+                            displayContent = phrase.title.formatWithInvisibleSymbolsPlain(singleLine: true)
+                        } else {
+                            displayContent = phrase.title.replacingNewlinesWithSpaces()
+                        }
                         if displayContent.count > 40 {
                             return String(displayContent.prefix(40)) + "..."
                         }
@@ -338,7 +344,12 @@ struct ClipHoldApp: App {
                     let displayText: String = {
                         let content = item.displayTitle
                         
-                        var displayContent = content.replacingNewlinesWithSpaces()
+                        var displayContent: String
+                        if showInvisibleCharacters {
+                            displayContent = content.formatWithInvisibleSymbolsPlain(singleLine: true)
+                        } else {
+                            displayContent = content.replacingNewlinesWithSpaces()
+                        }
                         let dateString = item.date.formatted(for: dateDisplayFormatInMenu, currentDate: dateReloader.now)
                         
                         let characterCountText = showCharacterCount ? String(localized:" - \(item.text.count)文字") : ""